app/requirements.txt
Django==2.2
django-crum==0.7.6
django-filter==2.2.0
djangorestframework==3.11.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2